✅ Complete popup_generators.py modularization
- Created map_locations/popup_generators.py with 6 functions
- Reduced core.py from 2,067 to 1,755 lines (-312 lines)
- Updated core.py to import from popup_generators module
- All backward compatibility maintained
- Testing: ✅ All imports and functionality work correctly

Progress: 67% Complete (4/6 modules)
Priority 2 Content modules: popup_generators.py ✅
Next: exporters.py (Priority 2: Medium Risk - depends on popup_generators)
